- fixed int/unsigned int

+++ Code/GraphMol/ROMol.cpp
@@ -591,10 +591,11 @@ unsigned int ROMol::addConformersFromTrajectory(const RDGeom::Trajectory *traj,
   PRECONDITION(traj, "traj must not be NULL");
   PRECONDITION(traj->numPoints() == this->getNumAtoms(),
                "Number of atom mismatch between ROMol and Trajectory");
-  PRECONDITION(nConf <= traj->size(), "nConf must be <= traj->size()");
+  PRECONDITION(nConf <= static_cast<int>(traj->size()),
+               "nConf must be <= traj->size()");
   if (nConf == -1)
     nConf = traj->size();
-  for (unsigned int n = 0; n < nConf; ++n) {
+  for (int n = 0; n < nConf; ++n) {
     Conformer *conf = new Conformer(this->getNumAtoms());
     for (unsigned int i = 0; i < this->getNumAtoms(); ++i)
       conf->setAtomPos(i, traj->getSnapshot(n).getPoint3D(i));
